+++ bpx/schema.py
@@ -1,11 +1,6 @@
-from typing import List, Literal, Union, Dict
-
-from dataclasses import dataclass
-
+from typing import List, Literal, Union, Dict, get_args
 from pydantic import BaseModel, Field, Extra, root_validator
-
 from bpx import Function, InterpolatedTable, check_sto_limits
-
 from warnings import warn
 
 FloatFunctionTable = Union[float, Function, InterpolatedTable]
@@ -15,7 +10,7 @@ class ExtraBaseModel(BaseModel):
     class Config:
         extra = Extra.forbid
 
-    @dataclass
+    # @dataclass
     class settings:
         v_tol: float = 0.001  # Absolute tolerance in [V] to validate the voltage limits
 
@@ -282,6 +277,30 @@ class ElectrodeBlendedSPM(ContactBase):
     particle: Dict[str, Particle] = Field(alias="Particle")
 
 
+class UserDefined(BaseModel):
+    class Config:
+        extra = Extra.allow
+
+    def __init__(self, **data):
+        """
+        Overwrite the default __init__ to convert strings to Function objects and
+        dicts to InterpolatedTable objects
+        """
+        for k, v in data.items():
+            if isinstance(v, str):
+                data[k] = Function(v)
+            elif isinstance(v, dict):
+                data[k] = InterpolatedTable(**v)
+        super().__init__(**data)
+
+    @root_validator(pre=True)
+    def validate_extra_fields(cls, values):
+        for k, v in values.items():
+            if not isinstance(v, get_args(FloatFunctionTable)):
+                raise TypeError(f"{k} must be of type 'FloatFunctionTable'")
+        return values
+
+
 class Experiment(ExtraBaseModel):
     time: List[float] = Field(
         alias="Time [s]",
@@ -322,6 +341,10 @@ class Parameterisation(ExtraBaseModel):
     separator: Contact = Field(
         alias="Separator",
     )
+    user_defined: UserDefined = Field(
+        None,
+        alias="User-defined",
+    )
 
     # Reusable validators
     _sto_limit_validation = root_validator(skip_on_failure=True, allow_reuse=True)(
@@ -338,6 +361,10 @@ class ParameterisationSPM(ExtraBaseModel):
     )
     positive_electrode: Union[ElectrodeSingleSPM, ElectrodeBlendedSPM] = Field(
         alias="Positive electrode",
+    )
+    user_defined: UserDefined = Field(
+        None,
+        alias="User-defined",
     )
 
     # Reusable validators